e2e tests: move whereabouts client to its own pkg
