+++ e2e/e2e_test.go
@@ -22,20 +22,17 @@ import (
 	"k8s.io/client-go/tools/clientcmd"
 
 	nettypes "github.com/k8snetworkplumbingwg/network-attachment-definition-client/pkg/apis/k8s.cni.cncf.io/v1"
-	netclient "github.com/k8snetworkplumbingwg/network-attachment-definition-client/pkg/client/clientset/versioned/typed/k8s.cni.cncf.io/v1"
 
+	wbtestclient "github.com/k8snetworkplumbingwg/whereabouts/e2e/client"
 	"github.com/k8snetworkplumbingwg/whereabouts/e2e/entities"
 	testenv "github.com/k8snetworkplumbingwg/whereabouts/e2e/testenvironment"
 	"github.com/k8snetworkplumbingwg/whereabouts/pkg/api/whereabouts.cni.cncf.io/v1alpha1"
-	wbclient "github.com/k8snetworkplumbingwg/whereabouts/pkg/client/clientset/versioned"
 	wbstorage "github.com/k8snetworkplumbingwg/whereabouts/pkg/storage/kubernetes"
 	"github.com/k8snetworkplumbingwg/whereabouts/pkg/types"
 )
 
 const (
-	createTimeout   = 10 * time.Second
-	deleteTimeout   = 2 * createTimeout
-	rsCreateTimeout = 600 * time.Second
+	createPodTimeout = 10 * time.Second
 )
 
 func TestWhereaboutsE2E(t *testing.T) {
@@ -54,7 +51,7 @@ var _ = Describe("Whereabouts functionality", func() {
 		)
 
 		var (
-			clientInfo   *ClientInfo
+			clientInfo   *wbtestclient.ClientInfo
 			testConfig   *testenv.Configuration
 			netAttachDef *nettypes.NetworkAttachmentDefinition
 			pod          *core.Pod
@@ -73,18 +70,18 @@ var _ = Describe("Whereabouts functionality", func() {
 			config, err = clusterConfig()
 			Expect(err).NotTo(HaveOccurred())
 
-			clientInfo, err = NewClientInfo(config)
+			clientInfo, err = wbtestclient.NewClientInfo(config)
 			Expect(err).NotTo(HaveOccurred())
 
 			netAttachDef = macvlanNetworkWithWhereaboutsIPAMNetwork(testNetworkName, testNamespace, ipv4TestRange)
 
 			By("creating a NetworkAttachmentDefinition for whereabouts")
-			_, err = clientInfo.addNetAttachDef(netAttachDef)
+			_, err = clientInfo.AddNetAttachDef(netAttachDef)
 			Expect(err).NotTo(HaveOccurred())
 		})
 
 		AfterEach(func() {
-			Expect(clientInfo.delNetAttachDef(netAttachDef)).To(Succeed())
+			Expect(clientInfo.DelNetAttachDef(netAttachDef)).To(Succeed())
 		})
 
 		Context("Single pod tests", func() {
@@ -93,18 +90,18 @@ var _ = Describe("Whereabouts functionality", func() {
 				var err error
 
 				By("creating a pod with whereabouts net-attach-def")
-				pod, err = clientInfo.provisionPod(
+				pod, err = clientInfo.ProvisionPod(
 					singlePodName,
 					testNamespace,
 					podTierLabel(singlePodName),
-					podNetworkSelectionElements(testNetworkName),
+					entities.PodNetworkSelectionElements(testNetworkName),
 				)
 				Expect(err).NotTo(HaveOccurred())
 			})
 
 			AfterEach(func() {
 				By("deleting pod with whereabouts net-attach-def")
-				Expect(clientInfo.deletePod(pod)).To(Succeed())
+				Expect(clientInfo.DeletePod(pod)).To(Succeed())
 			})
 
 			It("allocates a single pod within the correct IP range", func() {
@@ -135,12 +132,12 @@ var _ = Describe("Whereabouts functionality", func() {
 				}, ipPoolNamespace)
 				Expect(err).NotTo(HaveOccurred())
 
-				replicaSet, err = clientInfo.provisionReplicaSet(
+				replicaSet, err = clientInfo.ProvisionReplicaSet(
 					rsName,
 					testNamespace,
 					emptyReplicaSet,
 					podTierLabel(rsName),
-					podNetworkSelectionElements(testNetworkName),
+					entities.PodNetworkSelectionElements(testNetworkName),
 				)
 				Expect(err).NotTo(HaveOccurred())
 			})
@@ -152,7 +149,7 @@ var _ = Describe("Whereabouts functionality", func() {
 						clientInfo, k8sIPAM, rsName, testNamespace, ipPoolName, testNetworkName)).To(Succeed())
 
 				By("deleting replicaset with whereabouts net-attach-def")
-				Expect(clientInfo.deleteReplicaSet(replicaSet)).To(Succeed())
+				Expect(clientInfo.DeleteReplicaSet(replicaSet)).To(Succeed())
 			})
 
 			It("allocates each IP pool entry with a unique pod IP", func() {
@@ -165,19 +162,20 @@ var _ = Describe("Whereabouts functionality", func() {
 					allPods, err := clientInfo.Client.CoreV1().Pods(core.NamespaceAll).List(context.TODO(), metav1.ListOptions{})
 					Expect(err).NotTo(HaveOccurred())
 
-					replicaSet, err = clientInfo.updateReplicaSet(entities.ReplicaSetObject(
-						testConfig.MaxReplicas(allPods.Items),
-						rsName,
-						testNamespace,
-						podTierLabel(rsName),
-						podNetworkSelectionElements(testNetworkName),
-					))
+					replicaSet, err = clientInfo.UpdateReplicaSet(
+						entities.ReplicaSetObject(
+							testConfig.MaxReplicas(allPods.Items),
+							rsName,
+							testNamespace,
+							podTierLabel(rsName),
+							entities.PodNetworkSelectionElements(testNetworkName),
+						))
 					Expect(err).NotTo(HaveOccurred())
 					Expect(
-						WaitForReplicaSetSteadyState(
+						wbtestclient.WaitForReplicaSetSteadyState(
 							clientInfo.Client,
 							testNamespace,
-							podTierString(rsName),
+							entities.ReplicaSetQuery(rsName),
 							replicaSet,
 							rsSteadyTimeout)).To(Succeed())
 					Expect(
@@ -186,7 +184,7 @@ var _ = Describe("Whereabouts functionality", func() {
 							clientInfo.Client,
 							testNamespace,
 							ipPoolName,
-							podTierString(rsName))).To(Succeed())
+							entities.ReplicaSetQuery(rsName))).To(Succeed())
 				}
 			})
 		})
@@ -206,7 +204,7 @@ var _ = Describe("Whereabouts functionality", func() {
 			Context("regular sized network", func() {
 				BeforeEach(func() {
 					var err error
-					_, err = clientInfo.provisionStatefulSet(statefulSetName, namespace, serviceName, initialReplicaNumber, testNetworkName)
+					_, err = clientInfo.ProvisionStatefulSet(statefulSetName, namespace, serviceName, initialReplicaNumber, testNetworkName)
 					Expect(err).NotTo(HaveOccurred())
 					Expect(
 						clientInfo.Client.CoreV1().Pods(namespace).List(
@@ -215,7 +213,7 @@ var _ = Describe("Whereabouts functionality", func() {
 				})
 
 				AfterEach(func() {
-					Expect(clientInfo.deleteStatefulSet(namespace, serviceName, selector)).To(Succeed())
+					Expect(clientInfo.DeleteStatefulSet(namespace, serviceName, selector)).To(Succeed())
 					Expect(
 						clientInfo.Client.CoreV1().Pods(namespace).List(
 							context.TODO(), metav1.ListOptions{LabelSelector: selector})).To(
@@ -241,7 +239,7 @@ var _ = Describe("Whereabouts functionality", func() {
 				})
 
 				table.DescribeTable("stateful sets scale up / down", func(testSetup func(int), instanceDelta int) {
-					const scaleTimeout = createTimeout * 6
+					const scaleTimeout = createPodTimeout * 6
 
 					testSetup(instanceDelta)
 
@@ -257,28 +255,28 @@ var _ = Describe("Whereabouts functionality", func() {
 						HaveLen(initialReplicaNumber), "we should have one allocation for each live pod")
 				},
 					table.Entry("scale up then down 5 replicas", func(deltaInstances int) {
-						Expect(clientInfo.scaleStatefulSet(serviceName, namespace, deltaInstances)).To(Succeed())
-						Expect(clientInfo.scaleStatefulSet(serviceName, namespace, -deltaInstances)).To(Succeed())
+						Expect(clientInfo.ScaleStatefulSet(serviceName, namespace, deltaInstances)).To(Succeed())
+						Expect(clientInfo.ScaleStatefulSet(serviceName, namespace, -deltaInstances)).To(Succeed())
 					}, 5),
 					table.Entry("scale up then down 10 replicas", func(deltaInstances int) {
-						Expect(clientInfo.scaleStatefulSet(serviceName, namespace, deltaInstances)).To(Succeed())
-						Expect(clientInfo.scaleStatefulSet(serviceName, namespace, -deltaInstances)).To(Succeed())
+						Expect(clientInfo.ScaleStatefulSet(serviceName, namespace, deltaInstances)).To(Succeed())
+						Expect(clientInfo.ScaleStatefulSet(serviceName, namespace, -deltaInstances)).To(Succeed())
 					}, 10),
 					table.Entry("scale up then down 20 replicas", func(deltaInstances int) {
-						Expect(clientInfo.scaleStatefulSet(serviceName, namespace, deltaInstances)).To(Succeed())
-						Expect(clientInfo.scaleStatefulSet(serviceName, namespace, -deltaInstances)).To(Succeed())
+						Expect(clientInfo.ScaleStatefulSet(serviceName, namespace, deltaInstances)).To(Succeed())
+						Expect(clientInfo.ScaleStatefulSet(serviceName, namespace, -deltaInstances)).To(Succeed())
 					}, 20),
 					table.Entry("scale down then up 5 replicas", func(deltaInstances int) {
-						Expect(clientInfo.scaleStatefulSet(serviceName, namespace, -deltaInstances)).To(Succeed())
-						Expect(clientInfo.scaleStatefulSet(serviceName, namespace, deltaInstances)).To(Succeed())
+						Expect(clientInfo.ScaleStatefulSet(serviceName, namespace, -deltaInstances)).To(Succeed())
+						Expect(clientInfo.ScaleStatefulSet(serviceName, namespace, deltaInstances)).To(Succeed())
 					}, 5),
 					table.Entry("scale down then up 10 replicas", func(deltaInstances int) {
-						Expect(clientInfo.scaleStatefulSet(serviceName, namespace, -deltaInstances)).To(Succeed())
-						Expect(clientInfo.scaleStatefulSet(serviceName, namespace, deltaInstances)).To(Succeed())
+						Expect(clientInfo.ScaleStatefulSet(serviceName, namespace, -deltaInstances)).To(Succeed())
+						Expect(clientInfo.ScaleStatefulSet(serviceName, namespace, deltaInstances)).To(Succeed())
 					}, 10),
 					table.Entry("scale down then up 20 replicas", func(deltaInstances int) {
-						Expect(clientInfo.scaleStatefulSet(serviceName, namespace, -deltaInstances)).To(Succeed())
-						Expect(clientInfo.scaleStatefulSet(serviceName, namespace, deltaInstances)).To(Succeed())
+						Expect(clientInfo.ScaleStatefulSet(serviceName, namespace, -deltaInstances)).To(Succeed())
+						Expect(clientInfo.ScaleStatefulSet(serviceName, namespace, deltaInstances)).To(Succeed())
 					}, 20),
 				)
 			})
@@ -296,30 +294,30 @@ var _ = Describe("Whereabouts functionality", func() {
 
 				BeforeEach(func() {
 					var err error
-					tinyNetwork, err = clientInfo.addNetAttachDef(
+					tinyNetwork, err = clientInfo.AddNetAttachDef(
 						macvlanNetworkWithWhereaboutsIPAMNetwork(networkName, namespace, rangeWithTwoIPs))
 					Expect(err).NotTo(HaveOccurred())
 
-					_, err = clientInfo.provisionStatefulSet(statefulSetName, namespace, serviceName, replicaNumber, networkName)
+					_, err = clientInfo.ProvisionStatefulSet(statefulSetName, namespace, serviceName, replicaNumber, networkName)
 					Expect(err).NotTo(HaveOccurred())
 				})
 
 				AfterEach(func() {
-					Expect(clientInfo.delNetAttachDef(tinyNetwork)).To(Succeed())
-					Expect(clientInfo.deleteStatefulSet(namespace, serviceName, selector)).To(Succeed())
+					Expect(clientInfo.DelNetAttachDef(tinyNetwork)).To(Succeed())
+					Expect(clientInfo.DeleteStatefulSet(namespace, serviceName, selector)).To(Succeed())
 				})
 
 				It("IPPool is exhausted", func() {
 					const scaleUpReplicas = 1
-					Expect(clientInfo.scaleStatefulSet(serviceName, namespace, scaleUpReplicas)).To(Succeed())
+					Expect(clientInfo.ScaleStatefulSet(serviceName, namespace, scaleUpReplicas)).To(Succeed())
 					Expect(
-						WaitForStatefulSetCondition(
+						wbtestclient.WaitForStatefulSetCondition(
 							clientInfo.Client,
 							namespace,
 							serviceName,
 							replicaNumber+scaleUpReplicas,
 							statefulSetCreateTimeout,
-							isStatefulSetReadyPredicate)).To(HaveOccurred(), "the IPPool is already at its limits")
+							wbtestclient.IsStatefulSetReadyPredicate)).To(HaveOccurred(), "the IPPool is already at its limits")
 				})
 
 				Context("deleting a pod from the statefulset", func() {
@@ -347,22 +345,22 @@ var _ = Describe("Whereabouts functionality", func() {
 						Expect(clientInfo.Client.CoreV1().Pods(namespace).Delete(
 							context.TODO(), podName, metav1.DeleteOptions{GracePeriodSeconds: &rightNow})).To(Succeed())
 
-						Expect(WaitForStatefulSetCondition(
+						Expect(wbtestclient.WaitForStatefulSetCondition(
 							clientInfo.Client,
 							namespace,
 							serviceName,
 							replicaNumber,
 							time.Second,
-							isStatefulSetDegradedPredicate)).Should(Succeed())
+							wbtestclient.IsStatefulSetDegradedPredicate)).Should(Succeed())
 
-						scaleUpTimeout := 2 * createTimeout
-						Expect(WaitForStatefulSetCondition(
+						scaleUpTimeout := 2 * createPodTimeout
+						Expect(wbtestclient.WaitForStatefulSetCondition(
 							clientInfo.Client,
 							namespace,
 							serviceName,
 							replicaNumber,
 							scaleUpTimeout,
-							isStatefulSetReadyPredicate)).Should(Succeed())
+							wbtestclient.IsStatefulSetReadyPredicate)).Should(Succeed())
 					})
 
 					It("can recover from an exhausted IP pool", func() {
@@ -391,7 +389,7 @@ func iPPoolConsistency(k8sIPAM *wbstorage.KubernetesIPAM, cs *kubernetes.Clients
 	}
 
 	ipPoolAllocations := ipPool.Allocations()
-	podList, err := ListPods(cs, namespace, podLabel)
+	podList, err := wbtestclient.ListPods(cs, namespace, podLabel)
 	if err != nil {
 		return err
 	}
@@ -492,124 +490,8 @@ func podTierLabel(podTier string) map[string]string {
 	return map[string]string{tier: podTier}
 }
 
-func podNetworkSelectionElements(networkNames ...string) map[string]string {
-	return map[string]string{
-		nettypes.NetworkAttachmentAnnot: strings.Join(networkNames, ","),
-	}
-}
-
-type ClientInfo struct {
-	Client    *kubernetes.Clientset
-	NetClient netclient.K8sCniCncfIoV1Interface
-	WbClient  wbclient.Interface
-}
-
-func NewClientInfo(config *rest.Config) (*ClientInfo, error) {
-	clientSet, err := kubernetes.NewForConfig(config)
-	if err != nil {
-		return nil, err
-	}
-	netClient, err := netclient.NewForConfig(config)
-	if err != nil {
-		return nil, err
-	}
-
-	wbClient, err := wbclient.NewForConfig(config)
-	if err != nil {
-		return nil, err
-	}
-
-	return &ClientInfo{
-		Client:    clientSet,
-		NetClient: netClient,
-		WbClient:  wbClient,
-	}, nil
-}
-
-func (c *ClientInfo) addNetAttachDef(netattach *nettypes.NetworkAttachmentDefinition) (*nettypes.NetworkAttachmentDefinition, error) {
-	return c.NetClient.NetworkAttachmentDefinitions(netattach.ObjectMeta.Namespace).Create(context.TODO(), netattach, metav1.CreateOptions{})
-}
-
-func (c *ClientInfo) delNetAttachDef(netattach *nettypes.NetworkAttachmentDefinition) error {
-	return c.NetClient.NetworkAttachmentDefinitions(netattach.ObjectMeta.Namespace).Delete(context.TODO(), netattach.Name, metav1.DeleteOptions{})
-}
-
-func (c *ClientInfo) provisionPod(podName string, namespace string, label, annotations map[string]string) (*core.Pod, error) {
-	pod := entities.PodObject(podName, namespace, label, annotations)
-	pod, err := c.Client.CoreV1().Pods(pod.Namespace).Create(context.Background(), pod, metav1.CreateOptions{})
-	if err != nil {
-		return nil, err
-	}
-
-	const podCreateTimeout = 10 * time.Second
-	if err := WaitForPodReady(c.Client, pod.Namespace, pod.Name, podCreateTimeout); err != nil {
-		return nil, err
-	}
-
-	pod, err = c.Client.CoreV1().Pods(pod.Namespace).Get(context.Background(), pod.Name, metav1.GetOptions{})
-	if err != nil {
-		return nil, err
-	}
-
-	return pod, nil
-}
-
-func (c *ClientInfo) deletePod(pod *core.Pod) error {
-	if err := c.Client.CoreV1().Pods(pod.Namespace).Delete(context.TODO(), pod.Name, metav1.DeleteOptions{}); err != nil {
-		return err
-	}
-
-	const podDeleteTimeout = 20 * time.Second
-	if err := WaitForPodToDisappear(c.Client, pod.GetNamespace(), pod.GetName(), podDeleteTimeout); err != nil {
-		return err
-	}
-	return nil
-}
-
-func (c *ClientInfo) provisionReplicaSet(rsName string, namespace string, replicaCount int32, labels, annotations map[string]string) (*v1.ReplicaSet, error) {
-	replicaSet, err := c.Client.AppsV1().ReplicaSets(namespace).Create(
-		context.Background(),
-		entities.ReplicaSetObject(replicaCount, rsName, namespace, labels, annotations),
-		metav1.CreateOptions{})
-	if err != nil {
-		return nil, err
-	}
-
-	const rsCreateTimeout = 600 * time.Second
-	if err := WaitForPodBySelector(c.Client, namespace, podTierString(rsName), rsCreateTimeout); err != nil {
-		return nil, err
-	}
-
-	replicaSet, err = c.Client.AppsV1().ReplicaSets(namespace).Get(context.Background(), replicaSet.Name, metav1.GetOptions{})
-	if err != nil {
-		return nil, err
-	}
-
-	return replicaSet, nil
-}
-
-func (c *ClientInfo) updateReplicaSet(replicaSet *v1.ReplicaSet) (*v1.ReplicaSet, error) {
-	replicaSet, err := c.Client.AppsV1().ReplicaSets(replicaSet.GetNamespace()).Update(context.Background(), replicaSet, metav1.UpdateOptions{})
-	if err != nil {
-		return nil, err
-	}
-	return replicaSet, nil
-}
-
-func (c *ClientInfo) deleteReplicaSet(replicaSet *v1.ReplicaSet) error {
-	const rsDeleteTimeout = 2 * rsCreateTimeout
-	if err := c.Client.AppsV1().ReplicaSets(replicaSet.GetNamespace()).Delete(context.Background(), replicaSet.Name, metav1.DeleteOptions{}); err != nil {
-		return err
-	}
-
-	if err := WaitForReplicaSetToDisappear(c.Client, replicaSet.GetNamespace(), replicaSet.GetName(), rsDeleteTimeout); err != nil {
-		return err
-	}
-	return nil
-}
-
 // Waits for all replicas to be fully removed from replicaset, and checks that there are 0 ip pool allocations
-func checkZeroIPPoolAllocationsAndReplicas(clientInfo *ClientInfo, k8sIPAM *wbstorage.KubernetesIPAM, rsName, namespace string, ipPoolName string, networkNames ...string) error {
+func checkZeroIPPoolAllocationsAndReplicas(clientInfo *wbtestclient.ClientInfo, k8sIPAM *wbstorage.KubernetesIPAM, rsName, namespace string, ipPoolName string, networkNames ...string) error {
 	const (
 		emptyReplicaSet   = 0
 		rsSteadyTimeout   = 1200 * time.Second
@@ -617,20 +499,20 @@ func checkZeroIPPoolAllocationsAndReplicas(clientInfo *ClientInfo, k8sIPAM *wbst
 	)
 	var err error
 
-	replicaSet, err := clientInfo.updateReplicaSet(
+	replicaSet, err := clientInfo.UpdateReplicaSet(
 		entities.ReplicaSetObject(
 			emptyReplicaSet,
 			rsName,
 			namespace,
 			podTierLabel(rsName),
-			podNetworkSelectionElements(networkNames...),
+			entities.PodNetworkSelectionElements(networkNames...),
 		))
 	if err != nil {
 		return err
 	}
 
-	matchingLabel := podTierString(rsName)
-	if err = WaitForReplicaSetSteadyState(clientInfo.Client, namespace, matchingLabel, replicaSet, rsSteadyTimeout); err != nil {
+	matchingLabel := entities.ReplicaSetQuery(rsName)
+	if err = wbtestclient.WaitForReplicaSetSteadyState(clientInfo.Client, namespace, matchingLabel, replicaSet, rsSteadyTimeout); err != nil {
 		return err
 	}
 
@@ -638,60 +520,6 @@ func checkZeroIPPoolAllocationsAndReplicas(clientInfo *ClientInfo, k8sIPAM *wbst
 		return err
 	}
 
-	return nil
-}
-func (c *ClientInfo) provisionStatefulSet(statefulSetName string, namespace string, serviceName string, replicas int, networkNames ...string) (*v1.StatefulSet, error) {
-	const statefulSetCreateTimeout = 60 * createTimeout
-	statefulSet, err := c.Client.AppsV1().StatefulSets(namespace).Create(
-		context.TODO(),
-		entities.StatefulSetSpec(statefulSetName, namespace, serviceName, replicas, podNetworkSelectionElements(networkNames...)),
-		metav1.CreateOptions{})
-	if err != nil {
-		return nil, err
-	}
-
-	if err := WaitForStatefulSetCondition(
-		c.Client,
-		namespace,
-		serviceName,
-		replicas,
-		statefulSetCreateTimeout,
-		isStatefulSetReadyPredicate); err != nil {
-		return nil, err
-	}
-	return statefulSet, nil
-}
-
-func (c *ClientInfo) deleteStatefulSet(namespace string, serviceName string, labelSelector string) error {
-	const statefulSetDeleteTimeout = 6 * deleteTimeout
-
-	if err := c.Client.AppsV1().StatefulSets(namespace).Delete(
-		context.TODO(), serviceName, deleteRightNowAndBlockUntilAssociatedPodsAreGone()); err != nil {
-		return err
-	}
-
-	return WaitForStatefulSetGone(c.Client, namespace, serviceName, labelSelector, statefulSetDeleteTimeout)
-}
-
-func deleteRightNowAndBlockUntilAssociatedPodsAreGone() metav1.DeleteOptions {
-	var (
-		blockUntilAssociatedPodsAreGone = metav1.DeletePropagationForeground
-		rightNow                        = int64(0)
-	)
-	return metav1.DeleteOptions{GracePeriodSeconds: &rightNow, PropagationPolicy: &blockUntilAssociatedPodsAreGone}
-}
-
-func (c *ClientInfo) scaleStatefulSet(statefulSetName string, namespace string, deltaInstance int) error {
-	statefulSet, err := c.Client.AppsV1().StatefulSets(namespace).Get(context.TODO(), statefulSetName, metav1.GetOptions{})
-	if err != nil {
-		return err
-	}
-	newReplicas := *statefulSet.Spec.Replicas + int32(deltaInstance)
-	statefulSet.Spec.Replicas = &newReplicas
-
-	if _, err := c.Client.AppsV1().StatefulSets(namespace).Update(context.TODO(), statefulSet, metav1.UpdateOptions{}); err != nil {
-		return err
-	}
 	return nil
 }
 
@@ -779,8 +607,4 @@ func inRange(cidr string, ip string) error {
 	}
 
 	return fmt.Errorf("ip [%s] is NOT in range %s", ip, cidr)
-}
-
-func podTierString(rsName string) string {
-	return "tier=" + rsName
 }
